页面调整

+++ wms-service/src/main/java/com/wms/service/GspCustomerService.java
@@ -72,6 +72,7 @@ public class GspCustomerService extends BaseService {
 			gspCustomer.setIsCooperation(Constant.CODE_YES_OR_YES);
 			gspCustomer.setClientStartDate(DateUtil.parse(gspCustomerForm.getClientStartDate(),"yyyy-MM-dd"));
 			gspCustomer.setClientEndDate(DateUtil.parse(gspCustomerForm.getClientEndDate(),"yyyy-MM-dd"));
+			gspCustomer.setIsUse(Constant.IS_USE_YES);
 			gspCustomerMybatisDao.add(gspCustomer);
 
 			FirstReviewLogForm firstReviewLogForm = new FirstReviewLogForm();
@@ -95,8 +96,15 @@ public class GspCustomerService extends BaseService {
 	public Json editGspCustomer(GspCustomerForm gspCustomerForm) {
 		Json json = new Json();
 		GspCustomer gspCustomer = gspCustomerMybatisDao.queryById(gspCustomerForm.getClientId());
+		if(gspCustomer ==null){
+			return Json.error("没有查询到对应的申请单号");
+		}
+
+		if(!gspCustomer.getFirstState().equals(Constant.CODE_CATALOG_FIRSTSTATE_NEW)){
+			return Json.error("审核中的单据无法修改");
+		}
 		BeanUtils.copyProperties(gspCustomerForm, gspCustomer);
-		gspCustomerMybatisDao.update(gspCustomer);
+		gspCustomerMybatisDao.updateBySelective(gspCustomer);
 		json.setSuccess(true);
 		return json;
 	}
@@ -104,6 +112,13 @@ public class GspCustomerService extends BaseService {
 	public Json deleteGspCustomer(String id) {
 		Json json = new Json();
 		GspCustomer gspCustomer = gspCustomerMybatisDao.queryById(id);
+		if(gspCustomer ==null){
+			return Json.error("没有查询到对应的申请单号");
+		}
+
+		if(!gspCustomer.getFirstState().equals(Constant.CODE_CATALOG_FIRSTSTATE_NEW)){
+			return Json.error("审核中的单据无法修改");
+		}
 		if(gspCustomer != null){
 			gspCustomerMybatisDao.delete(gspCustomer);
 		}
@@ -157,6 +172,52 @@ public class GspCustomerService extends BaseService {
 				firstReviewLogService.updateFirstReviewByNo(s,Constant.CODE_CATALOG_CHECKSTATE_QCCHECKING);
 			}
 			return Json.success("确认成功");
+		}catch (Exception e){
+			e.printStackTrace();
+			TransactionAspectSupport.currentTransactionStatus().setRollbackOnly();
+			return Json.error("操作失败");
+		}
+
+	}
+
+	public Json reApply(String id){
+		try{
+			if(StringUtils.isEmpty(id)){
+				return Json.error("请选择要重新申请的数据");
+			}
+			String[] arr = id.split(",");
+			for(String s : arr){
+				String no = commonService.generateSeq(Constant.APLCUSNO,SfcUserLoginUtil.getLoginUser().getWarehouse().getId());
+				//更新就数据状态
+				GspCustomer gspCustomer = gspCustomerMybatisDao.queryById(s);
+				String oldNo = gspCustomer.getClientId();
+				gspCustomer.setFirstState(Constant.CODE_CATALOG_FIRSTSTATE_USELESS);
+				gspCustomer.setIsUse(Constant.IS_USE_NO);
+				gspCustomerMybatisDao.updateBySelective(gspCustomer);
+				firstReviewLogService.updateFirstReviewByNo(s,Constant.CODE_CATALOG_CHECKSTATE_FAIL);
+
+				//更新审核状态为未通过
+				gspCustomer.setClientId(no);
+				gspCustomer.setCreateId(SfcUserLoginUtil.getLoginUser().getId());
+				gspCustomer.setFirstState(Constant.CODE_CATALOG_FIRSTSTATE_NEW);
+				gspCustomer.setIsCheck(Constant.CODE_YES_OR_YES);
+				gspCustomer.setIsCooperation(Constant.CODE_YES_OR_YES);
+				gspCustomer.setIsUse(Constant.IS_USE_NO);
+				gspCustomerMybatisDao.add(gspCustomer);
+
+				//插入新的日志
+				FirstReviewLogForm firstReviewLogForm = new FirstReviewLogForm();
+				firstReviewLogForm.setCreateId(SfcUserLoginUtil.getLoginUser().getId());
+				firstReviewLogForm.setReviewTypeId(no);
+				firstReviewLogForm.setReviewId(RandomUtil.getUUID());
+				firstReviewLogForm.setApplyState(Constant.CODE_CATALOG_CHECKSTATE_NEW);
+				firstReviewLogService.addFirstReviewLog(firstReviewLogForm);
+
+				//TODO 审核通过的数据需要更新为失败(待测试)
+				firstReviewLogService.cancelData(oldNo);
+
+			}
+			return Json.success("重新申请成功成功");
 		}catch (Exception e){
 			e.printStackTrace();
 			TransactionAspectSupport.currentTransactionStatus().setRollbackOnly();
